Move independent modules out of cnd directory
Both, the action_to_http_request and the rfc7807 middleware are
modules independent from cnd. They can easily exist on a higher
level.
